pyproject.toml
[project]
name = "ai-travel-concierge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-cosmos>=4.14.4",
    "azure-identity>=1.25.1",
    "pydantic>=2.11.10",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "semantic-kernel>=1.36.0",
    "tiktoken>=0.12.0",
]
